Make update_counter generic for all split view
