+++ config/routes.rb
@@ -561,12 +561,15 @@ Rails.application.routes.draw do
                as: :work_package_progress
     end
 
+    get "/split_view/update_counter" => "work_packages/split_view#update_counter",
+        on: :member
+
     # states managed by client-side (angular) routing on work_package#show
     get "/" => "work_packages#index", on: :collection, as: "index"
     get "/create_new" => "work_packages#index", on: :collection, as: "new_split"
     get "/new" => "work_packages#index", on: :collection, as: "new", state: "new"
     # We do not want to match the work package export routes
-    get "(/*state)" => "work_packages#show", on: :member, as: "", constraints: { id: /\d+/, state: /(?!shares).+/ }
+    get "(/*state)" => "work_packages#show", on: :member, as: "", constraints: { id: /\d+/, state: /(?!(shares|split_view)).+/ }
     get "/share_upsale" => "work_packages#index", on: :collection, as: "share_upsale"
     get "/edit" => "work_packages#show", on: :member, as: "edit"
   end
@@ -708,9 +711,6 @@ Rails.application.routes.draw do
 
     get "/:work_package_id/close",
         action: :close_split_view
-
-    get "/:work_package_id/update_counter",
-        action: :update_counter
   end
 
   resources :notifications, only: :index do
